Refactor API tests setUp in a separate class
* Introduces a base class ApiTestCase for api test cases
* Renames ApiTestCase into FakeApiTestCase, as it actually tests fake
api
